+++ src/base/b-virtual-scroll/test/runners/init-load.js
@@ -1,0 +1,64 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+const
+	path = require('upath'),
+	pzlr = require('@pzlr/build-core'),
+	delay = require('delay');
+
+const
+	componentDir = pzlr.resolve.blockSync('b-virtual-scroll'),
+	h = require(path.join(componentDir, 'test/helpers.js'));
+
+module.exports = async (p, {componentSelector, component: c, components}) => {
+	describe('b-virtual-scroll', () => {
+		it('reloads data with initLoad', async () => {
+			const chunkSize = await c.evaluate((ctx) => ctx.chunkSize);
+
+			c.evaluate((ctx) => ctx.initLoad());
+			await h.waitForRefDisplay(p, {componentSelector}, 'tombstones', '');
+			await h.waitForRefDisplay(p, {componentSelector}, 'tombstones', 'none');
+
+			expect(await c.evaluate((ctx) => ctx.$refs.tombstones.style.display)).toBe('none');
+			expect(await c.evaluate((ctx) => ctx.$refs.container.style.display)).toBe('');
+			expect(await c.evaluate((ctx) => ctx.$refs.container.childElementCount)).toBe(chunkSize);
+		});
+
+		const
+			secondComponent = components[1],
+			secondComponentSelector = `${componentSelector}#second`;
+
+		const
+			selectors = {componentSelector: secondComponentSelector, componentName: componentSelector};
+
+		const
+			reloadCount = 6;
+
+		it('reloads data with frequent initLoad calls and renders correct data chunk to the page', async () => {
+			let i = reloadCount;
+
+			while (i--) {
+				secondComponent.evaluate((ctx) => ctx.initLoad());
+				await delay(100);
+				await h.waitForRefDisplay(p, selectors, 'tombstones', '');
+			}
+
+			const
+				chunkSize = await secondComponent.evaluate((ctx) => ctx.chunkSize),
+				elIndex = reloadCount * chunkSize;
+
+			await h.waitForRefDisplay(p, selectors, 'tombstones', 'none');
+			expect(await secondComponent.evaluate((ctx) => ctx.$refs.container.childElementCount)).toBe(chunkSize);
+			expect(await secondComponent.evaluate((ctx) => ctx.$refs.container.childElementCount)).toBe(chunkSize);
+
+			expect(
+				await secondComponent.evaluate((ctx) => ctx.$refs.container.children[0].getAttribute('data-index')
+			)).toBe(String(elIndex));
+		});
+	});
+};