+++ tests/cases.js
@@ -14,6 +14,8 @@ module.exports = [
 
 	'--name b-virtual-scroll --suit slot-empty --runner slot-empty',
 
+	'--name b-virtual-scroll --suit double-components --runner init-load',
+
 	// b-button
 	`--name b-button --suit demo`
 ];