+++ src/base/b-virtual-scroll/CHANGELOG.md
@@ -11,10 +11,18 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
+## v3.0.0-rc.25 (2020-06-03)
+
+#### :bug: Bug Fix
+
+* [Fixed an issue where skeletons disappeared](https://github.com/V4Fire/Client/issues/230)
+* [Fixed an issue with a race condition `chunk-request/init`](https://github.com/V4Fire/Client/issues/203)
+* [Fixed an issue where an `empty` slot appeared when there was data](https://github.com/V4Fire/Client/issues/259)
+
 ## v3.0.0-rc.19 (2020-05-26)
 
 #### :bug: Bug Fix
 
-* [Fixed rendering of truncated data in `b-virtual-scroll`](https://github.com/V4Fire/Client/issues/231)
-* [Fixed rendering of empty slot in `b-virtual-scroll`](https://github.com/V4Fire/Client/issues/241)
-* [Fixed clear of `in-view` in `b-virtual-scroll`](https://github.com/V4Fire/Client/pull/201)
+* [Fixed rendering of truncated data](https://github.com/V4Fire/Client/issues/231)
+* [Fixed rendering of empty slot](https://github.com/V4Fire/Client/issues/241)
+* [Fixed clear of `in-view`](https://github.com/V4Fire/Client/pull/201)